% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{analyze_sparsity_rpsde}
\alias{analyze_sparsity_rpsde}
\title{Analyze sparsity of matrices in the rSPDE approach}
\usage{
analyze_sparsity_rpsde(
  nu_lower_bound,
  nu_upper_bound,
  d,
  C,
  G,
  C_inv,
  C_inv_G,
  rspde_order
)
}
\arguments{
\item{nu_lower_bound}{Lower bound for the smoothness parameter}

\item{nu_upper_bound}{Upper bound for the smoothness parameter}

\item{d}{The dimension of the domain}

\item{C}{The mass matrix}

\item{G}{The stiffness matrix}

\item{C_inv}{Inverse of the mass matrix}

\item{C_inv_G}{C^{-1}G}

\item{rspde_order}{The order of the rational approximation}
}
\value{
A list containing informations on sparsity of the precision matrices
}
\description{
Auxiliar function to analyze sparsity of matrices in the rSPDE approach
}
