% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{analyze_sparsity_rspde}
\alias{analyze_sparsity_rspde}
\title{Analyze sparsity of matrices in the rSPDE approach}
\usage{
analyze_sparsity_rspde(
  nu_upper_bound,
  d,
  rspde_order,
  fem_mesh_matrices,
  include_lower_order = TRUE,
  include_higher_order = TRUE
)
}
\arguments{
\item{nu_upper_bound}{Upper bound for the smoothness parameter}

\item{d}{The dimension of the domain}

\item{rspde_order}{The order of the rational approximation}

\item{fem_mesh_matrices}{A list containing FEM-related matrices. The list should contain elements c0, g1, g2, g3, etc.}

\item{include_lower_order}{Logical. Should the lower-order terms be included? They are needed for the cases
when alpha = nu + d/2 is integer or for when sharp is set to TRUE.}

\item{include_higher_order}{Logical. Should be included for when nu is estimated or for when alpha = nu + d/2 is not an integer.}
}
\value{
A list containing informations on sparsity of the precision matrices
}
\description{
Auxiliar function to analyze sparsity of matrices in the rSPDE approach
}
