% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde_intrinsic.R
\name{bru_get_mapper.inla_rspde_fintrinsic}
\alias{bru_get_mapper.inla_rspde_fintrinsic}
\title{rSPDE inlabru mapper}
\usage{
\method{bru_get_mapper}{inla_rspde_fintrinsic}(model, ...)
}
\arguments{
\item{model}{An \code{inla_rspde_fintrinsic} object for which to construct or extract a mapper}

\item{\dots}{Arguments passed on to other methods}
}
\description{
rSPDE inlabru mapper
}
