% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde_intrinsic.R
\name{bru_get_mapper.intrinsic_matern}
\alias{bru_get_mapper.intrinsic_matern}
\alias{ibm_n.bru_mapper_intrinsic_matern}
\alias{ibm_values.bru_mapper_intrinsic_matern}
\alias{ibm_jacobian.bru_mapper_intrinsic_matern}
\title{rSPDE inlabru mapper}
\usage{
\method{bru_get_mapper}{intrinsic_matern}(model, ...)

\method{ibm_n}{bru_mapper_intrinsic_matern}(mapper, ...)

\method{ibm_values}{bru_mapper_intrinsic_matern}(mapper, ...)

\method{ibm_jacobian}{bru_mapper_intrinsic_matern}(mapper, input, ...)
}
\arguments{
\item{model}{An \code{intrinsic_matern} object for which to construct or extract a mapper}

\item{\dots}{Arguments passed on to other methods}

\item{mapper}{A \code{bru_mapper_intrinsic_matern} object}

\item{input}{The values for which to produce a mapping matrix}
}
\description{
rSPDE inlabru mapper
}
