% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlabru_rspde.R
\name{bru_mapper.inla.rspde}
\alias{bru_mapper.inla.rspde}
\alias{ibm_n.bru_mapper_inla_rspde}
\alias{ibm_values.bru_mapper_inla_rspde}
\alias{ibm_amatrix.bru_mapper_inla_rspde}
\title{rSPDE inlabru mapper}
\usage{
\method{bru_mapper}{inla.rspde}(model, ...)

\method{ibm_n}{bru_mapper_inla_rspde}(mapper, ...)

\method{ibm_values}{bru_mapper_inla_rspde}(mapper, ...)

\method{ibm_amatrix}{bru_mapper_inla_rspde}(mapper, input, ...)
}
\arguments{
\item{model}{An `inla.mesh.1d` or `inla.mesh.2d` object to use as a mapper}

\item{input}{The values for which to produce a mapping matrix}

\item{indexed}{logical; If `TRUE`, the `ibm_values()` output will be the
integer indexing sequence for the latent variables. If `FALSE`, the knot
locations are returned (useful as an interpolator for `rw2` models
and similar).
Default: `TRUE`}
}
\description{
rSPDE inlabru mapper
}
