% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{build_sparse_matrix_rspde}
\alias{build_sparse_matrix_rspde}
\title{Create sparse matrix from entries and graph}
\usage{
build_sparse_matrix_rspde(entries, graph)
}
\arguments{
\item{entries}{The entries of the precision matrix}

\item{graph}{The sparsity graph of the precision matrix}
}
\value{
index for rSPDE models.
}
\description{
Create sparse matrix from entries and graph
}
