% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional.computations.R
\name{CBrSPDE.matern.loglike}
\alias{CBrSPDE.matern.loglike}
\title{Log-likelihood function for latent Gaussian fractional SPDE model}
\usage{
CBrSPDE.matern.loglike(
  object,
  Y,
  A,
  sigma.e,
  mu = 0,
  user_nu = NULL,
  user_kappa = NULL,
  user_tau = NULL,
  user_m = NULL,
  pivot = TRUE
)
}
\arguments{
\item{object}{A cov_rSPDE object}

\item{Y}{The observations, either a vector or a matrix where
the columns correspond to independent replicates of observations.}

\item{A}{An observation matrix that links the measurement location to the finite element basis.}

\item{sigma.e}{The standard deviation of the measurement noise.}

\item{mu}{Expectation vector of the latent field (default = 0).}

\item{user_nu}{If non-null, update the shape parameter of the covariance function.}

\item{user_kappa}{If non-null, update the range parameter of the covariance function.}

\item{user_tau}{If non-null, update the standard deviation of the covariance function.}

\item{user_m}{If non-null, update the order of the rational approximation, which needs to be a positive integer.}

\item{pivot}{Should pivoting be used for the Cholesky decompositions? Default is TRUE}
}
\value{
A cov_rSPDE object
}
\description{
This function evaluates the log-likelihood function for covariance-based
rational approximation of a fractional SPDE model
}
