% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional.computations.R
\name{CBrSPDE.matern.loglike2}
\alias{CBrSPDE.matern.loglike2}
\title{Log-likelihood function for latent Gaussian fractional SPDE model}
\usage{
CBrSPDE.matern.loglike2(
  kappa,
  tau,
  nu,
  sigma.e,
  mu = 0,
  Y,
  G,
  C,
  A,
  d = 2,
  m = 2,
  pivot = TRUE
)
}
\arguments{
\item{kappa}{Range parameter of the latent process.}

\item{tau}{Standard deviation of the latent process.}

\item{nu}{Shape parameter of the latent process.}

\item{sigma.e}{The standard deviation of the measurement noise.}

\item{mu}{Expectation vector of the latent field (default = 0).}

\item{Y}{The observations, either a vector or a matrix where
the columns correspond to independent replicates of observations.}

\item{G}{The stiffness matrix of a finite element discretization of the domain.}

\item{C}{The mass matrix of a finite element discretization of the domain.}

\item{A}{A matrix linking the measurement locations to the basis of the FEM approximation of the latent model.}

\item{d}{The dimension of the domain. The default value is 2.}

\item{m}{The order of the rational approximation, which needs to be a positive integer.
The default value is 2.}

\item{pivot}{Should pivoting be used for the Cholesky decompositions? Default is TRUE}
}
\value{
A cov_rSPDE object
}
\description{
This function evaluates the log-likelihood function for covariance-based
rational approximation of a fractional SPDE model
}
