% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_class_inla_rspde}
\alias{check_class_inla_rspde}
\title{Check if the object inherits from inla.rspde class}
\usage{
check_class_inla_rspde(model)
}
\arguments{
\item{model}{A model to test if it inherits from inla.rspde}
}
\value{
Gives an error if the object does not inherit from inla.rspde
}
\description{
Check if the object inherits from inla.rspde class
}
