% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional.operators.R
\name{cov_function_mesh}
\alias{cov_function_mesh}
\title{Covariance between mesh nodes and locations}
\usage{
cov_function_mesh(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic for computing the covariance between mesh nodes and locations for
model objects.
}
