% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional.computations.R
\name{cov_rSPDE_matern.loglike}
\alias{cov_rSPDE_matern.loglike}
\alias{matern.cov_rSPDE.loglike}
\title{Log-likelihood function for latent Gaussian fractional SPDE model}
\usage{
cov_rSPDE_matern.loglike(
  object,
  Y,
  A,
  sigma.e,
  mu = 0,
  user_nu = NULL,
  user_kappa = NULL,
  user_tau = NULL,
  user_m = NULL,
  pivot = TRUE
)

matern.cov_rSPDE.loglike(
  kappa,
  tau,
  nu,
  sigma.e,
  mu = 0,
  Y,
  G,
  C,
  A,
  d = 2,
  m = 2,
  pivot = TRUE
)
}
\arguments{
\item{object}{A cov_rSPDE object}

\item{Y}{The observations, either a vector or a matrix where
the columns correspond to independent replicates of observations.}

\item{A}{A matrix linking the measurement locations to the basis of the FEM approximation of the latent model.}

\item{sigma.e}{The standard deviation of the measurement noise.}

\item{mu}{Expectation vector of the latent field (default = 0).}

\item{user_nu}{If non-null, update the shape parameter of the covariance function.}

\item{user_kappa}{If non-null, update the range parameter of the covariance function.}

\item{user_tau}{If non-null, update the standard deviation of the covariance function.}

\item{user_m}{If non-null, update the order of the rational approximation, which needs to be a positive integer.}

\item{pivot}{Should pivoting be used for the Cholesky decompositions? Default is TRUE}

\item{kappa}{Range parameter of the latent process.}

\item{tau}{Standard deviation of the latent process.}

\item{nu}{Shape parameter of the latent process.}

\item{G}{The stiffness matrix of a finite element discretization of the domain.}

\item{C}{The mass matrix of a finite element discretization of the domain.}

\item{d}{The dimension of the domain. The default value is 2.}

\item{m}{The order of the rational approximation, which needs to be a positive integer.
The default value is 2.}
}
\value{
A cov_rSPDE object

A cov_rSPDE object
}
\description{
This function evaluates the log-likelihood function for covariance-based
rational approximation of a fractional SPDE model

This function evaluates the log-likelihood function for covariance-based
rational approximation of a fractional SPDE model
}
