% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional.operators.R
\name{covariance_mesh}
\alias{covariance_mesh}
\title{Covariance between mesh nodes}
\usage{
covariance_mesh(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic for computing the covariance between mesh nodes for model objects.
}
