% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{create_rspde_model}
\alias{create_rspde_model}
\title{Create INLA-based rSPDE models}
\usage{
create_rspde_model(
  inla_mesh,
  nu_lower_bound = get_inla_mesh_dimension(inla_mesh)/4,
  nu_upper_bound = 4,
  rspde_order = 2,
  nu = NULL,
  sharp = TRUE,
  debug = FALSE,
  optimize = TRUE,
  prior.kappa = list(meanlog = 0, sdlog = 1),
  prior.nu = list(meanlog = 0, sdlog = 1),
  prior.tau = list(meanlog = 0, sdlog = 1)
)
}
\arguments{
\item{inla_mesh}{An INLA mesh}

\item{nu_lower_bound}{Lower bound for the smoothness parameter.}

\item{nu_upper_bound}{Upper bound for the smoothness parameter.}

\item{rspde_order}{The order of the covariance-based rational SPDE approach.}

\item{nu}{If nu is set to a parameter, nu will be kept fixed and will not be estimated.
If nu is NULL, it will be estimated.}

\item{sharp}{The sparsity graph should have the correct sparsity (costs more to perform
a sparsity analysis) or an upper bound for the sparsity? If TRUE, the graph
will have the correct sparsity.}

\item{debug}{INLA debug argument.}

\item{optimize}{Should the model be optimized? In this case the sparsities of
the matrices will be analyzed.}

\item{prior.kappa}{a list containing the elements meanlog and sdlog, that is,
the mean and standard deviation on the log scale.}

\item{prior.nu}{a list containing the elements meanlog and sdlog, that is,
the mean and standard deviation on the log scale.}

\item{prior.tau}{a list containing the elements meanlog and sdlog, that is,
the mean and standard deviation on the log scale.}
}
\value{
An INLA model.
}
\description{
Create INLA-based rSPDE models with general smoothness
}
