% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create_summary_from_density}
\alias{create_summary_from_density}
\title{Creates a summary from a density data frame}
\usage{
create_summary_from_density(density_df, name)
}
\arguments{
\item{density_df}{A density data frame}

\item{name}{Name of the parameter}
}
\value{
A data frame containing a basic summary
}
\description{
Auxiliar function to create summaries from density data drames
}
