% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cut_decimals}
\alias{cut_decimals}
\title{Approximation function for covariance-based rSPDE models}
\usage{
cut_decimals(nu)
}
\arguments{
\item{nu}{A real number}
}
\value{
An approximation
}
\description{
Approximation function to be used to compute the precision matrix for covariance-based rSPDE models
}
