% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_inla_mesh_dimension}
\alias{get_inla_mesh_dimension}
\title{Get the dimension of an INLA mesh}
\usage{
get_inla_mesh_dimension(inla_mesh)
}
\arguments{
\item{inla_mesh}{An INLA mesh}
}
\value{
The dimension of an INLA mesh.
}
\description{
Get the dimension of an INLA mesh
}
