% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{gg_df}
\alias{gg_df}
\title{Data frame for result objects from R-INLA fitted models to be used in ggplot2}
\usage{
gg_df(result, ...)
}
\arguments{
\item{result}{a result object for which the data frame is desired}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A data frame containing the posterior densities.
}
\description{
Data frame for result objects from R-INLA fitted models to be used in ggplot2
}
