% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{gg_df.rspde_result}
\alias{gg_df.rspde_result}
\title{Data frame for rspde_result objects to be used in ggplot2}
\usage{
\method{gg_df}{rspde_result}(
  result,
  parameter = result$params,
  transform = TRUE,
  restrict_x_axis = parameter,
  restrict_quantiles = list(std.dev = c(0, 1), range = c(0, 1), nu = c(0, 1), kappa =
    c(0, 1), tau = c(0, 1)),
  ...
)
}
\arguments{
\item{result}{An rspde_result object.}

\item{parameter}{Vector. Which parameters to get the posterior density in the data.frame? The options are \code{std.dev}, \code{range}, \code{tau}, \code{kappa} and \code{nu}.}

\item{transform}{Should the posterior density be given in the original scale?}

\item{restrict_x_axis}{Variables to restrict the range of x axis based on quantiles.}

\item{restrict_quantiles}{List of quantiles to restrict x axis.}

\item{...}{currently not used.}
}
\value{
A data frame containing the posterior densities.
}
\description{
Returns a ggplot-friendly data-frame with the marginal posterior densities.
}
