% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{graph_repl_rspde}
\alias{graph_repl_rspde}
\title{Extraction of vector of replicates for 'INLA'}
\usage{
graph_repl_rspde(graph_spde, repl = NULL, group = NULL, group_col = NULL)
}
\arguments{
\item{graph_spde}{An \code{rspde_metric_graph} object built with the
\code{rspde.metric_graph()} function from the 'rSPDE' package.}

\item{repl}{Which replicates? If there is no replicates, one
can set \code{repl} to \code{NULL}. If one wants all replicates,
then one sets to \code{repl} to \verb{__all}.}

\item{group}{Which groups? If there is no groups, one
can set \code{group} to \code{NULL}. If one wants all groups,
then one sets to \code{group} to \verb{__all}.}

\item{group_col}{Which "column" of the data contains the group variable?}
}
\value{
The vector of replicates paired with groups.
}
\description{
Extracts the vector of replicates from an 'rSPDE'
model object for 'INLA'
}
