% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{likelihood_process_parameters}
\alias{likelihood_process_parameters}
\title{Process parameters likelihood}
\usage{
likelihood_process_parameters(theta, param_vector, which_par, logscale)
}
\arguments{
\item{theta}{vector of parameters}

\item{param_vector}{vector of parameters to be used}

\item{which_par}{which parameter to consider}

\item{logscale}{log scale?}
}
\value{
The value in the correct scale
}
\description{
Process parameters likelihood
}
