% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional.operators.R
\name{make_A}
\alias{make_A}
\title{Projection matrix for model objects}
\usage{
make_A(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic for computing the projection matrix that links observation locations
to model mesh nodes.
}
