% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional.operators.R
\name{matern.cov_rSPDE.operators}
\alias{matern.cov_rSPDE.operators}
\title{Create INLA-based rSPDE models}
\usage{
matern.cov_rSPDE.operators(C, G, nu, kappa, tau, m = 2, d)
}
\arguments{
\item{C}{The mass matrix of a finite element discretization of the domain of interest.}

\item{G}{The stiffness matrix of a finite element discretization of the domain of interest.}

\item{nu}{Shape parameter of the covariance function.}

\item{kappa}{Range parameter of the covariance function.}

\item{tau}{Standard deviation of the covariance function.}

\item{m}{The order of the rational approximation, which needs to be a positive integer.
The default value is 2.}

\item{d}{The dimension of the domain.}
}
\value{
A cov_rSPDE object.
}
\description{
\code{matern.operators} is used for computing a
covariance-based rational SPDE approximation of a stationary Gaussian random
fields on \eqn{R^d} with a Matern covariance function
\deqn{C(h) = \frac{\tau^2}{2^(\nu-1)\Gamma(\nu)}(\kappa h)^\nu K_\nu(\kappa h)}{C(h) =
(\tau^2/(2^(\nu-1)\Gamma(\nu))(\kappa h)^\nu K_\nu(\kappa h)}
}
