% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary1d.R
\name{matern.rational.precision}
\alias{matern.rational.precision}
\title{Precision matrix of rational approximation of Matern covariance}
\usage{
matern.rational.precision(
  loc,
  order,
  nu,
  kappa,
  sigma,
  type_rational = "brasil",
  type_interp = "spline",
  equally_spaced = FALSE,
  ordering = c("field", "location")
)
}
\arguments{
\item{loc}{Locations at which the precision is evaluated}

\item{order}{Order of the rational approximation}

\item{nu}{Smoothness parameter}

\item{kappa}{Range parameter}

\item{sigma}{Standard deviation}

\item{type_rational}{Method used to compute the coefficients of the rational approximation.}

\item{type_interp}{Interpolation method for the rational coefficients.}

\item{equally_spaced}{Are the locations equally spaced?}

\item{ordering}{Return the matrices ordered by field or by location?}
}
\value{
A list containing the precision matrix \code{Q} of the process and its derivatives if they exist, and
a matrix \code{A} that extracts the elements corresponding to the process.
}
\description{
Computes the precision matrix for a rational approximation of the Matern covariance on intervals.
}
\examples{
h <- seq(from = 0, to = 1, length.out = 100)
cov.true <- matern.covariance(h, kappa = 10, sigma = 1, nu = 0.8)
Q.approx <- matern.rational.precision(h, kappa = 10, sigma = 1, nu = 0.8, order = 2)
cov.approx <- Q.approx$A\%*\%solve(Q.approx$Q,Q.approx$A[1,])

plot(h, cov.true)
lines(h, cov.approx, col = 2)
}
