% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{posterior_df}
\alias{posterior_df}
\title{Create a posterior data frame for rspde.result objects}
\usage{
posterior_df(
  rspde_result,
  parameter = c("tau", "kappa", "nu"),
  transform = TRUE
)
}
\arguments{
\item{rspde_result}{An rspde.result object.}

\item{parameter}{Vector. Which parameters to get the posterior density in the data.frame? The options are \code{tau}, \code{kappa} and \code{nu}.}

\item{transform}{Should the posterior density be given in the original scale?}
}
\value{
A data frame containing the posterior densities.
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
