% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{precision.inla_rspde}
\alias{precision.inla_rspde}
\title{Get the precision matrix of \code{inla_rspde} objects}
\usage{
\method{precision}{inla_rspde}(object, theta = NULL, ...)
}
\arguments{
\item{object}{The \code{inla_rspde} object obtained with the \code{rspde.matern()} function.}

\item{theta}{If null, the starting values for theta will be used. Otherwise, it must be suplied as a vector.
For stationary models, we have \code{theta = c(log(tau), log(kappa), nu)}. For nonstationary models, we have
\code{theta = c(theta_1, theta_2, ..., theta_n, nu)}.}

\item{...}{Currently not used.}
}
\value{
The precision matrix.
}
\description{
Function to get the precision matrix of an \code{inla_rspde} object created with the \code{rspde.matern()} function.
}
\seealso{
\code{\link[=precision.CBrSPDEobj]{precision.CBrSPDEobj()}}, \code{\link[=matern.operators]{matern.operators()}}
}
