% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rational.order<-}
\alias{rational.order<-}
\title{Changing the order of the rational approximation}
\usage{
rational.order(x) <- value
}
\arguments{
\item{x}{A \code{CBrSPDE} or an \code{rpsde.inla} object}

\item{value}{The order of rational approximation.}
}
\value{
An object of the same class with the new order
of rational approximation.
}
\description{
Changing the order of the rational approximation
}
