% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rational.order}
\alias{rational.order}
\title{Get the order of rational approximation.}
\usage{
rational.order(object)
}
\arguments{
\item{object}{A \code{CBrSPDEobj} object or an \code{inla.rspde} object.}
}
\value{
The order of rational approximation.
}
\description{
Get the order of rational approximation.
}
