% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rational.type<-}
\alias{rational.type<-}
\title{Change the type of the rational approximation}
\usage{
rational.type(x) <- value
}
\arguments{
\item{x}{A \code{CBrSPDE} or an \code{rpsde.inla} object}

\item{value}{The type of rational approximation. The current options are "chebfun", "brasil" and "chebfunLB"}
}
\value{
An object of the same class with the new rational approximation.
}
\description{
Change the type of the rational approximation
}
