% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rational.type}
\alias{rational.type}
\title{Get type of rational approximation.}
\usage{
rational.type(object)
}
\arguments{
\item{object}{A \code{CBrSPDEobj} object or an \code{inla.rspde} object.}
}
\value{
The type of rational approximation.
}
\description{
Get type of rational approximation.
}
