% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{rspde.inla}
\alias{rspde.inla}
\title{rSPDE wrapper to run INLA}
\usage{
rspde.inla(..., n.tries = 10)
}
\arguments{
\item{...}{Arguments for `inla` function.}

\item{n.tries}{The number of tries}
}
\description{
Wrapper to run INLA
}
