% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{rspde.matern.precision.opt}
\alias{rspde.matern.precision.opt}
\title{Optimized precision matrix of the covariance-based rational
approximation}
\usage{
rspde.matern.precision.opt(
  kappa,
  nu,
  tau,
  rspde.order,
  dim,
  fem_matrices,
  graph = NULL,
  sharp,
  type_rational_approx,
  wl2_table = NULL
)
}
\arguments{
\item{kappa}{Range parameter of the covariance function.}

\item{nu}{Shape parameter of the covariance function.}

\item{tau}{Scale parameter of the covariance function.}

\item{rspde.order}{The order of the rational approximation}

\item{dim}{The dimension of the domain}

\item{fem_matrices}{A list containing the FEM-related matrices.
The list should contain elements C, G, G_2, G_3, etc.}

\item{graph}{The sparsity graph of the matrices. If NULL, only a vector
of the elements will be returned, if non-NULL, a sparse matrix will
be returned.}

\item{sharp}{The sparsity graph should have the correct sparsity (costs
more to perform a sparsity analysis) or an upper bound for the sparsity?}

\item{type_rational_approx}{Which type of rational approximation
should be used? The current types are "brasil", "chebfun", "chebfunLB" or
"wl2".}

\item{wl2_table}{A table of weighted-L2 coefficients, only used for
\code{type_rational_approx = "wl2"}. If \code{NULL}, it is taken from the tables
stored in the package where there is one for this configuration, and
computed otherwise.}
}
\value{
The precision matrix
}
\description{
\code{rspde.matern.precision} is used for computing the
optimized version of the precision matrix of the
covariance-based rational SPDE approximation of a stationary Gaussian random
fields on \eqn{R^d} with a Matern covariance function
\deqn{C(h) = \frac{\sigma^2}{2^{\nu-1}\Gamma(\nu)}(\kappa h)^\nu
K_\nu(\kappa h).}
}
