% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde_cgeneric.R
\name{rspde.matern_cgeneric}
\alias{rspde.matern_cgeneric}
\title{Matern rSPDE model object for INLA}
\usage{
rspde.matern_cgeneric(
  mesh,
  nu_upper_bound = 4,
  rspde_order = 2,
  nu = NULL,
  sharp = TRUE,
  debug = FALSE,
  optimize = TRUE,
  prior.kappa = NULL,
  prior.nu = NULL,
  prior.tau = NULL,
  start.lkappa = NULL,
  start.nu = NULL,
  start.ltau = NULL,
  prior.nu.dist = c("beta", "lognormal"),
  nu.prec.inc = 1,
  type.rational.approx = c("chebfun", "brasil", "chebfunLB")
)
}
\arguments{
\item{mesh}{The mesh to build the model. Should be an \code{inla.mesh} or
an \code{inla.mesh.1d} object.}

\item{nu_upper_bound}{Upper bound for the smoothness parameter.}

\item{rspde_order}{The order of the covariance-based rational SPDE approach.}

\item{nu}{If nu is set to a parameter, nu will be kept fixed and will not
be estimated. If nu is \code{NULL}, it will be estimated.}

\item{sharp}{The sparsity graph should have the correct sparsity (costs
more to perform a sparsity analysis) or an upper bound for the sparsity? If
\code{TRUE}, the graph will have the correct sparsity.}

\item{debug}{INLA debug argument.}

\item{optimize}{Should the model be optimized? In this case the sparsities of
the matrices will be analyzed.}

\item{prior.kappa}{a \code{list} containing the elements \code{meanlog} and
\code{sdlog}, that is, the mean and standard deviation on the log scale.}

\item{prior.nu}{a list containing the elements \code{mean} and \code{prec}
for beta distribution, or \code{loglocation} and \code{logscale} for a
truncated lognormal distribution. \code{loglocation} stands for
the location parameter of the truncated lognormal distribution in the log
scale. \code{prec} stands for the precision of a beta distribution.
\code{logscale} stands for the scale of the truncated lognormal
distribution on the log scale. Check details below.}

\item{prior.tau}{a list containing the elements \code{meanlog} and
\code{sdlog}, that is, the mean and standard deviation on the log scale.}

\item{start.lkappa}{Starting value for log of kappa.}

\item{start.nu}{Starting value for nu.}

\item{start.ltau}{Starting value for log of tau.}

\item{prior.nu.dist}{The distribution of the smoothness parameter.
The current options are "beta" or "lognormal". The default is "beta".}

\item{nu.prec.inc}{Amount to increase the precision in the beta prior
distribution. Check details below.}

\item{type.rational.approx}{Which type of rational approximation
should be used? The current types are "chebfun", "brasil" or "chebfunLB".}
}
\value{
An INLA model.
}
\description{
Creates an INLA object for a stationary Matern model with
general smoothness parameter.
}
