% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{rspde.precision}
\alias{rspde.precision}
\title{Precision matrices for rSPDE models Calculates the precision matrix 
for given parameter values based on an rspde model object.}
\usage{
rspde.precision(rspde, theta, optimized = FALSE)
}
\arguments{
\item{rspde}{An rspde object.}

\item{theta}{Vector of log(kappa), log(nu)-dim/2, log(tau).}

\item{optimized}{Logical indicating if only the elements of the precision
matrix should be returned.}
}
\value{
Precision matrix for the rspde model.
}
\description{
Precision matrices for rSPDE models
}
