% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{rspde_make_index}
\alias{rspde_make_index}
\title{Create index for INLA-based rSPDE models}
\usage{
rspde_make_index(name, inla_mesh, rspde_order = 2, nu = NULL)
}
\arguments{
\item{name}{Name}

\item{inla_mesh}{An INLA mesh}

\item{rspde_order}{The order of the covariance-based rational SPDE approach.}

\item{nu}{If NULL, then the model will assume that nu will be estimated. If
nu is fixed, you should provide the value of nu.}
}
\value{
index for rSPDE models.
}
\description{
Create index for INLA-based rSPDE models
}
