% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{rspde_Prec}
\alias{rspde_Prec}
\title{Precision matrix of the covariance-based rational approximation}
\usage{
rspde_Prec(C, G, C_inv_G, C_inv, n_m, d, theta, epsilon_identity = 1e-05)
}
\arguments{
\item{C}{The mass matrix}

\item{G}{The stiffness matrix}

\item{C_inv_G}{C^{-1}G}

\item{C_inv}{The inverse of the mass matrix}

\item{n_m}{The order of the rational approximation.}

\item{d}{The dimension}

\item{theta}{vector containing the parameters of the model.}

\item{epsilon_identity}{a multiple of the identity matrix to be summed to 
the precision matrix.}
}
\value{
The precision matrix
}
\description{
Computes the precision matrix for the covariance-based rational SPDE
}
