% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{rspde_Prec_int}
\alias{rspde_Prec_int}
\title{Precision matrix of the covariance-based rational approximation}
\usage{
rspde_Prec_int(C, G, C_inv_G, C_inv, d, theta, nu, epsilon_identity = 1e-05)
}
\arguments{
\item{C}{The mass matrix}

\item{G}{The stiffness matrix}

\item{C_inv_G}{C^{-1}G}

\item{C_inv}{The inverse of the mass matrix}

\item{d}{The dimension of the domain}

\item{theta}{vector containing the parameters of the model.}

\item{nu}{the smoothness parameter}

\item{epsilon_identity}{a multiple of the identity matrix to be summed to 
the precision matrix.}
}
\value{
The precision matrix
}
\description{
Computes the precision matrix for the covariance-based rational SPDE
}
