% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{rspde_Prec_opt_int}
\alias{rspde_Prec_opt_int}
\title{Precision matrix of the covariance-based rational approximation}
\usage{
rspde_Prec_opt_int(kappa, nu, tau, d, fem_matrices, graph = NULL)
}
\arguments{
\item{kappa}{Range parameter of the latent process.}

\item{nu}{The smoothness parameter}

\item{tau}{The precision parameter}

\item{d}{The dimension}

\item{fem_matrices}{A list containing the FEM-related matrices. The list should contain elements C, G, G_2, G_3, etc.}

\item{graph}{The sparsity graph of the matrices. If NULL, only a vector
of the elements will be returned, if non-NULL, a sparse matrix will be returned.}
}
\value{
The precision matrix
}
\description{
Computes the precision matrix for the covariance-based rational SPDE
}
