% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional.computations.R
\name{simulate.CBrSPDEobj}
\alias{simulate.CBrSPDEobj}
\title{Simulation of a fractional SPDE using a rational SPDE approximation}
\usage{
\method{simulate}{CBrSPDEobj}(
  object,
  nsim = 1,
  user_nu = NULL,
  user_kappa = NULL,
  user_tau = NULL,
  user_m = NULL,
  pivot = TRUE,
  ...
)
}
\arguments{
\item{object}{A CBrSPDE object}

\item{nsim}{The number of simulations.}

\item{user_nu}{If non-null, update the shape parameter of the covariance function.}

\item{user_kappa}{If non-null, update the range parameter of the covariance function.}

\item{user_tau}{If non-null, update the standard deviation of the covariance function.}

\item{user_m}{If non-null, update the order of the rational approximation, which needs to be a positive integer.}

\item{pivot}{Should pivoting be used for the Cholesky decompositions? Default is TRUE}

\item{...}{Currently not used.}
}
\value{
A CBrSPDE object
}
\description{
The function samples a Gaussian random field based on a pre-computed rational SPDE approximation.
}
