% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{'inla.rgeneric.cov_rspde_general'}
\alias{'inla.rgeneric.cov_rspde_general'}
\alias{inla.rgeneric.cov_rspde_general}
\title{Generic INLA method for the covariance-based rSPDE approach}
\usage{
inla.rgeneric.cov_rspde_general(
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  theta = NULL,
  args = NULL,
  ...
)
}
\arguments{
\item{cmd}{INLA parameter}

\item{theta}{Parameters of the model}

\item{args}{Arguments.}

\item{...}{Additional arguments to the call, or arguments with changed values. Use name = NULL to remove the argument name.}
}
\value{
An INLA model
}
\description{
Generic INLA method for the covariance-based rSPDE approach
}
