% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.R
\name{summary.CBrSPDEobj2d}
\alias{summary.CBrSPDEobj2d}
\alias{print.summary.CBrSPDEobj2d}
\alias{print.CBrSPDEobj2d}
\title{Summarise CBrSPDEobj2d objects}
\usage{
\method{summary}{CBrSPDEobj2d}(object, ...)

\method{print}{summary.CBrSPDEobj2d}(x, ...)

\method{print}{CBrSPDEobj2d}(x, ...)
}
\arguments{
\item{object}{an object of class "CBrSPDEobj2d", usually, a result of a call
to \code{\link[=matern2d.operators]{matern2d.operators()}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class "summary.CBrSPDEobj2d", usually, a result of a call
to \code{\link[=summary.CBrSPDEobj2d]{summary.CBrSPDEobj2d()}}.}
}
\description{
Summary method for class "CBrSPDEobj2d"
}
\examples{
n_loc <- 2000
loc_2d_mesh <- matrix(runif(n_loc * 2), n_loc, 2)
mesh_2d <- fm_mesh_2d(loc = loc_2d_mesh, cutoff = 0.03, max.edge = c(0.1, 0.5))
op <- matern2d.operators(mesh = mesh_2d)
op
}
