% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.R
\name{summary.intrinsicCBrSPDEobj}
\alias{summary.intrinsicCBrSPDEobj}
\alias{print.summary.intrinsicCBrSPDEobj}
\alias{print.intrinsicCBrSPDEobj}
\title{Summary method for class "intrinsicCBrSPDEobj"}
\usage{
\method{summary}{intrinsicCBrSPDEobj}(object, ...)

\method{print}{summary.intrinsicCBrSPDEobj}(x, ...)

\method{print}{intrinsicCBrSPDEobj}(x, ...)
}
\arguments{
\item{object}{an object of class "intrinsicCBrSPDEobj", usually, a result of a call
to \code{\link[=intrinsic.operators]{intrinsic.operators()}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class "summary.intrinsicCBrSPDEobj", usually, a result of a call
to \code{\link[=summary.intrinsicCBrSPDEobj]{summary.intrinsicCBrSPDEobj()}}.}
}
\description{
Summary method for class "intrinsicCBrSPDEobj"
}
