% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{summary.rspde.result}
\alias{summary.rspde.result}
\title{Summary for posteriors of rSPDE field parameters}
\usage{
\method{summary}{rspde.result}(object, digits = 6, ...)
}
\arguments{
\item{object}{A rspde.result object.}

\item{digits}{integer, used for number formatting with signif()}

\item{...}{Currently not used.}
}
\value{
Returns a data frame
containing the summary.
}
\description{
Summary for posteriors of rSPDE field parameters
}
