% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspde_lme.R
\name{summary.rspde_lme}
\alias{summary.rspde_lme}
\title{Summary Method for \code{rspde_lme} Objects.}
\usage{
\method{summary}{rspde_lme}(object, all_times = FALSE, ...)
}
\arguments{
\item{object}{an object of class "rspde_lme" containing results from the fitted model.}

\item{...}{not used.}
}
\value{
An object of class \code{summary_rspde_lme} containing several
informations of a \emph{rspde_lme} object.
}
\description{
Function providing a summary of results related to mixed effects regression models with Whittle-Matern latent models.
}
