% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.R
\name{summary.rSPDEobj}
\alias{summary.rSPDEobj}
\alias{print.summary.rSPDEobj}
\alias{print.rSPDEobj}
\title{Summarise rSPDE objects}
\usage{
\method{summary}{rSPDEobj}(object, ...)

\method{print}{summary.rSPDEobj}(x, ...)

\method{print}{rSPDEobj}(x, ...)
}
\arguments{
\item{object}{an object of class "rSPDEobj", usually, a result of a call
to \code{\link[=fractional.operators]{fractional.operators()}}, \code{\link[=matern.operators]{matern.operators()}}, or
\code{\link[=spde.matern.operators]{spde.matern.operators()}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class "summary.rSPDEobj", usually, a result of a call
to \code{\link[=summary.rSPDEobj]{summary.rSPDEobj()}}.}
}
\description{
Summary method for class "rSPDEobj"
}
