% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.R
\name{summary.rSPDEobj1d}
\alias{summary.rSPDEobj1d}
\alias{print.summary.rSPDEobj1d}
\alias{print.rSPDEobj1d}
\title{Summarise rSPDE objects without FEM}
\usage{
\method{summary}{rSPDEobj1d}(object, ...)

\method{print}{summary.rSPDEobj1d}(x, ...)

\method{print}{rSPDEobj1d}(x, ...)
}
\arguments{
\item{object}{an object of class "rSPDEobj1d", usually, a result of a call
to \code{\link[=matern.rational]{matern.rational()}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class "summary.rSPDEobj1d", usually, a result of a call
to \code{\link[=summary.rSPDEobj1d]{summary.rSPDEobj1d()}}.}
}
\description{
Summary method for class "rSPDEobj1d"
}
