% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.R
\name{summary.spacetimeobj}
\alias{summary.spacetimeobj}
\alias{print.summary.spacetimeobj}
\alias{print.spacetimeobj}
\title{Summarise spacetime objects}
\usage{
\method{summary}{spacetimeobj}(object, ...)

\method{print}{summary.spacetimeobj}(x, ...)

\method{print}{spacetimeobj}(x, ...)
}
\arguments{
\item{object}{an object of class "spacetimeobj", usually, a result of a call
to \code{\link[=spacetime.operators]{spacetime.operators()}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class "summary.spacetimeobj", usually, a result of a call
to \code{\link[=summary.spacetimeobj]{summary.spacetimeobj()}}.}
}
\description{
Summary method for class "spacetimeobj"
}
