% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional.computations.R
\name{update.cov_rSPDEobj}
\alias{update.cov_rSPDEobj}
\title{Update parameters of cov_rSPDE objects}
\usage{
\method{update}{cov_rSPDEobj}(
  object,
  user_nu = NULL,
  user_kappa = NULL,
  user_tau = NULL,
  user_m = NULL,
  ...
)
}
\arguments{
\item{object}{A cov_rSPDE object}

\item{user_nu}{If non-null, update the shape parameter of the covariance function.}

\item{user_kappa}{If non-null, update the range parameter of the covariance function.}

\item{user_tau}{If non-null, update the standard deviation of the covariance function.}

\item{user_m}{If non-null, update the order of the rational approximation, which needs to be a positive integer.}

\item{...}{Currently not used.}
}
\value{
A cov_rSPDE object
}
\description{
Function to change the parameters of a cov_rSPDE object
}
