% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.spde.R
\name{update.hybrid_spde}
\alias{update.hybrid_spde}
\title{Update parameters of hybrid_spde objects}
\usage{
\method{update}{hybrid_spde}(object, X = NULL, beta_X = NULL, ...)
}
\arguments{
\item{object}{A \code{hybrid_spde} object.}

\item{X}{If non-null, update the covariate matrix.}

\item{beta_X}{If non-null, update the regression coefficients.}

\item{...}{Additional arguments passed to the underlying update
method of \code{\link[=spde.matern.operators]{spde.matern.operators()}}.}
}
\value{
An object of class \code{"hybrid_spde"} with the updated
parameters.
}
\description{
Updates the model parameters of a \code{hybrid_spde} object.
Delegates the update of the covariance part to the underlying
\code{\link[=spde.matern.operators]{spde.matern.operators()}} update method and additionally rebuilds the
operator used to evaluate the mean and recomputes the stored mean
vector.
}
\seealso{
\code{\link[=hybrid.spde]{hybrid.spde()}}
}
