% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde_anisotropic.R
\name{bru_get_mapper.inla_rspde_anisotropic2d}
\alias{bru_get_mapper.inla_rspde_anisotropic2d}
\title{rSPDE anisotropic inlabru mapper}
\usage{
bru_get_mapper.inla_rspde_anisotropic2d(model, ...)
}
\arguments{
\item{model}{An \code{inla_rspde_anisotropic2d} object for which to construct or extract a mapper}

\item{\dots}{Arguments passed on to other methods}
}
\description{
rSPDE anisotropic inlabru mapper
}
