% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde_spacetime.R
\name{bru_get_mapper.inla_rspde_spacetime}
\alias{bru_get_mapper.inla_rspde_spacetime}
\title{rSPDE space time inlabru mapper}
\usage{
\method{bru_get_mapper}{inla_rspde_spacetime}(model, ...)
}
\arguments{
\item{model}{An \code{inla_rspde_spacetime} object for which to construct or extract a mapper}

\item{\dots}{Arguments passed on to other methods}
}
\description{
rSPDE space time inlabru mapper
}
