% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlabru_rspde.R
\name{bru_mapper.inla.rspde}
\alias{bru_mapper.inla.rspde}
\alias{bru_mapper.inla_rspde}
\alias{ibm_n.bru_mapper_inla_rspde}
\alias{ibm_values.bru_mapper_inla_rspde}
\alias{ibm_amatrix.bru_mapper_inla_rspde}
\alias{bru_get_mapper.inla_rspde}
\title{rSPDE inlabru mapper}
\usage{
\method{bru_mapper}{inla_rspde}(model, ...)

\method{ibm_n}{bru_mapper_inla_rspde}(mapper, ...)

\method{ibm_values}{bru_mapper_inla_rspde}(mapper, ...)

\method{ibm_amatrix}{bru_mapper_inla_rspde}(mapper, input, ...)

bru_get_mapper.inla_rspde(model, ...)
}
\arguments{
\item{model}{An `inla_rspde` object to use as a mapper}

\item{input}{The values for which to produce a mapping matrix}

\item{x}{The model to be passed to obtain the mapper.}
}
\description{
rSPDE inlabru mapper
}
