% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlabru_rspde.R
\name{bru_mapper.inla_rspde}
\alias{bru_mapper.inla_rspde}
\alias{ibm_n.bru_mapper_inla_rspde}
\alias{ibm_values.bru_mapper_inla_rspde}
\alias{ibm_amatrix.bru_mapper_inla_rspde}
\alias{bru_get_mapper.inla_rspde}
\title{rSPDE inlabru mapper}
\usage{
bru_mapper.inla_rspde(model, ...)

ibm_n.bru_mapper_inla_rspde(mapper, ...)

ibm_values.bru_mapper_inla_rspde(mapper, ...)

ibm_amatrix.bru_mapper_inla_rspde(mapper, input, ...)

bru_get_mapper.inla_rspde(model, ...)
}
\arguments{
\item{model}{An \code{inla_rspde} for which to construct or extract a mapper}

\item{\dots}{Arguments passed on to other methods}

\item{mapper}{A \code{bru_mapper.inla_rspde} object}

\item{input}{The values for which to produce a mapping matrix}
}
\description{
rSPDE inlabru mapper
}
