% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{create_rspde_model}
\alias{create_rspde_model}
\title{Create INLA-based rSPDE models}
\usage{
create_rspde_model(
  inla_mesh,
  nu_lower_bound = get_inla_mesh_dimension(inla_mesh)/4,
  nu_upper_bound = 7,
  rspde_order = 2,
  nu = NULL,
  debug = FALSE,
  optimize = FALSE,
  controls = list()
)
}
\arguments{
\item{inla_mesh}{An INLA mesh}

\item{nu_lower_bound}{Lower bound for the smoothness parameter.}

\item{nu_upper_bound}{Upper bound for the smoothness parameter.}

\item{rspde_order}{The order of the covariance-based rational SPDE approach.}

\item{nu}{If nu is set to a parameter, nu will be kept fixed and will not be estimated.
If nu is NULL, it will be estimated.}

\item{debug}{INLA debug argument.}

\item{optimize}{Should the model be optimized? In this case the sparsities of
the matrices will be analyzed.}

\item{controls}{A list of controls. For now, the only control is
sum.identity, which sums a multiple of the identity matrix (the value of 
the multiple is indicated by the user) to the
precision parameter to improve numerical stability.}
}
\value{
An INLA model.
}
\description{
Create INLA-based rSPDE models with general smoothness
}
