% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlabru_rspde.R
\name{cross_validation}
\alias{cross_validation}
\title{Perform cross-validation on a list of fitted models.}
\usage{
cross_validation(
  models,
  model_names = NULL,
  scores = c("mse", "crps", "scrps", "dss"),
  cv_type = c("k-fold", "loo", "lpo"),
  data = NULL,
  k = 5,
  percentage = 30,
  number_folds = 10,
  n_samples = 1000,
  return_scores_folds = FALSE,
  true_CV = FALSE,
  print = TRUE,
  fit_verbose = FALSE
)
}
\arguments{
\item{models}{A fitted model obtained from calling the \code{bru()} function or a list of models fitted with the \code{bru()} function.}

\item{model_names}{A vector containing the names of the models to appear in the returned \code{data.frame}. If \code{NULL}, the names will be of the form \verb{Model 1}, \verb{Model 2}, and so on.}

\item{scores}{A vector containing the scores to be computed. The options are "mse", "crps", "scrps" and "dss". By default, all scores are computed.}

\item{cv_type}{The type of the folding to be carried out. The options are \code{k-fold} for \code{k}-fold cross-validation, in which case the parameter \code{k} should be provided,
\code{loo}, for leave-one-out and \code{lpo} for leave-percentage-out, in this case, the parameter \code{percentage} should be given, and also the \code{number_folds}
with the number of folds to be done. The default is \code{k-fold}.}

\item{data}{A dataset to be used to perform crossvalidation on all models. Observe that this means that the data must contain all the variables for all the models. If \code{NULL},
then the data will be extracted from the first fitted model. Observe that if \code{data} is \code{NULL}, then the models need to be trained on the same dataset as the first model of the list.}

\item{k}{The number of folds to be used in \code{k}-fold cross-validation. Will only be used if \code{cv_type} is \code{k-fold}.}

\item{percentage}{The percentage (from 1 to 99) of the data to be used to train the model. Will only be used if \code{cv_type} is \code{lpo}.}

\item{number_folds}{Number of folds to be done if \code{cv_type} is \code{lpo}.}

\item{n_samples}{Number of samples to compute the posterior statistics to be used to compute the scores.}

\item{return_scores_folds}{If \code{TRUE}, the scores for each fold will also be returned.}

\item{true_CV}{Should a \code{TRUE} cross-validation be performed? If \code{TRUE} the models will be fitted on the training dataset. If \code{FALSE}, the parameters will be kept fixed at the ones obtained in the result object.}

\item{print}{Should partial results be printed throughout the computation?}

\item{fit_verbose}{Should INLA's run during cross-validation be verbose?}
}
\value{
A data.frame with the fitted models and the corresponding scores.
}
\description{
Obtain several scores for a list of fitted models according
to a folding scheme.
}
