% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{gg_df}
\alias{gg_df}
\title{Data frame for rspde.result objects to be used in ggplot2}
\usage{
gg_df(
  rspde_result,
  parameter = rspde_result$params,
  transform = TRUE,
  restrict_x_axis = parameter,
  restrict_quantiles = list(std.dev = c(0, 1), range = c(0, 1), nu = c(0, 1), kappa =
    c(0, 1), tau = c(0, 1))
)
}
\arguments{
\item{rspde_result}{An rspde.result object.}

\item{parameter}{Vector. Which parameters to get the posterior density in the data.frame? The options are \code{std.dev}, \code{range}, \code{tau}, \code{kappa} and \code{nu}.}

\item{transform}{Should the posterior density be given in the original scale?}
<<<<<<< HEAD

\item{restrict_x_axis}{Variables to restrict the range of x axis based on quantiles.}

\item{restrict_quantiles}{List of quantiles to restrict x axis.}
=======
>>>>>>> 3df757e7767cd6e85be3a9b38f2335b3985c1f9e
}
\value{
A data frame containing the posterior densities.
}
\description{
Returns a ggplot-friendly data-frame with the marginal posterior densities.
}
\examples{
# ADD_EXAMPLES_HERE
}
