% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspde_lme.R
\name{glance.rspde_lme}
\alias{glance.rspde_lme}
\alias{glance}
\title{Glance at an \code{rspde_lme} object}
\usage{
\method{glance}{rspde_lme}(x, ...)
}
\arguments{
\item{x}{An \code{rspde_lme} object.}

\item{...}{Currently not used.}
}
\value{
A \code{\link[tidyr:reexports]{tidyr::tibble()}} with exactly one row and columns:
\itemize{
\item \code{nobs} Number of observations used.
\item \code{sigma} the square root of the estimated residual variance
\item \code{logLik} The log-likelihood of the model.
\item \code{AIC} Akaike's Information Criterion for the model.
\item \code{BIC} Bayesian Information Criterion for the model.
\item \code{deviance} Deviance of the model.
\item \code{df.residual} Residual degrees of freedom.
\item \code{model.type} Type of latent model fitted.
}
}
\description{
Glance accepts a \code{rspde_lme} object and returns a
\code{\link[tidyr:reexports]{tidyr::tibble()}} with exactly one row of model summaries.
The summaries are the square root of the estimated variance of the measurement error, residual
degrees of freedom, AIC, BIC, log-likelihood,
the type of latent model used in the fit and the total number of observations.
}
\seealso{
\link{augment.rspde_lme}
}
