% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{plot.rspde.result}
\alias{plot.rspde.result}
\title{Posterior plots for rSPDE field parameters}
\usage{
\method{plot}{rspde.result}(
  x,
  which = c("tau", "kappa", "nu"),
  caption = list("Posterior density for tau", "Posterior density for kappa",
    "Posterior density for nu"),
  sub.caption = NULL,
  type_plot = "l",
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  main = "",
  cex.oma.main = 1.25,
  cex.caption = 1,
  ylab = "Density",
  xlab = "x",
  ...
)
}
\arguments{
\item{x}{A rspde.result object.}

\item{which}{For which parameters the posterior should be plotted?}

\item{caption}{captions to appear above the plots; character vector or list of valid graphics annotations. Can be set to "" or NA to suppress all captions.}

\item{sub.caption}{common title-above the figures if there are more than one.}

\item{type_plot}{what type of plot should be drawn. The default is 'l'.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot.}

\item{main}{character; title to be placed at each plot additionally (and above) all captions.}

\item{cex.oma.main}{controls the size of the sub.caption only if that is above the figures when there is more than one.}

\item{cex.caption}{controls the size of caption.}

\item{ylab}{Label for y axis.}

\item{xlab}{Label for x axis.}

\item{...}{Additional arguments.}
}
\value{
Called for its side effects.
}
\description{
Posterior plots for rSPDE field parameters
}
