% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional.computations.R
\name{predict.CBrSPDEobj}
\alias{predict.CBrSPDEobj}
\title{Prediction of a fractional SPDE using the covariance-based rational SPDE approximation}
\usage{
\method{predict}{CBrSPDEobj}(
  object,
  A,
  Aprd,
  Y,
  sigma.e,
  mu = 0,
  compute.variances = FALSE,
  pivot = TRUE,
  ...
)
}
\arguments{
\item{object}{A cov_rSPDE object}

\item{A}{A matrix linking the measurement locations to the basis of the FEM approximation of the latent model.}

\item{Aprd}{A matrix linking the prediction locations to the basis of the FEM approximation of the latent model.}

\item{Y}{A vector with the observed data, can also be a matrix where the columns are observations
of independent replicates of \eqn{u}.}

\item{sigma.e}{The standard deviation of the Gaussian measurement noise. Put to zero if the model
does not have measurement noise.}

\item{mu}{Expectation vector of the latent field (default = 0).}

\item{compute.variances}{Set to also TRUE to compute the kriging variances.}

\item{pivot}{Should pivoting be used on the Cholesky decompositions?}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list with elements
\item{mean }{The kriging predictor (the posterior mean of u|Y).}
\item{variance }{The posterior variances (if computed).}
}
\description{
The function is used for computing kriging predictions based on data \eqn{Y_i = u(s_i) + \epsilon_i},
where \eqn{\epsilon}{\epsilon} is mean-zero Gaussian measurement noise and \eqn{u(s)}{u(s)} is defined by
a fractional SPDE \eqn{L^\beta u(s) = W}{L^\beta u(s) = W}, where \eqn{W}{W} is Gaussian white noise.
}
