% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{rspde.make.A}
\alias{rspde.make.A}
\title{Create A matrices for rSPDE models}
\usage{
rspde.make.A(
  mesh = NULL,
  A = NULL,
  dim = NULL,
  loc = NULL,
  rspde_order = 2,
  nu = NULL,
  index = NULL,
  group = NULL,
  repl = 1L,
  n.group = NULL,
  n.repl = NULL
)
}
\arguments{
\item{mesh}{An INLA mesh, optional}

\item{A}{The A matrix from the standard SPDE approach. Should only be provided if an
INLA mesh is not provided.}

\item{dim}{the dimension. Should only be provided if an
INLA mesh is not provided.}

\item{loc}{Locations, needed if an INLA mesh is provided}

\item{rspde_order}{The order of the covariance-based rational SPDE approach.}

\item{nu}{If NULL, then the model will assume that nu will be estimated. If
nu is fixed, you should provide the value of nu.}

\item{index}{For each observation/prediction value, an index into loc. Default is seq_len(nrow(A.loc)).}

\item{group}{For each observation/prediction value, an index into the group model.}

\item{repl}{For each observation/prediction value, the replicate index.}

\item{n.group}{The size of the group model.}

\item{n.repl}{The total number of replicates.}
}
\value{
The A matrix for rSPDE models.
}
\description{
Create A matrices for INLA-based rSPDE models
}
