% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{rspde.make.index}
\alias{rspde.make.index}
\title{Create index for INLA-based rSPDE models}
\usage{
rspde.make.index(
  name,
  n.spde = NULL,
  n.group = 1,
  n.repl = 1,
  mesh = NULL,
  rspde_order = 2,
  nu = NULL,
  dim = NULL
)
}
\arguments{
\item{name}{Name}

\item{n.spde}{The number of basis functions in the mesh model.}

\item{n.group}{The size of the group model.}

\item{n.repl}{The total number of replicates.}

\item{mesh}{An INLA mesh}

\item{rspde_order}{The order of the covariance-based rational SPDE approach.}

\item{nu}{If NULL, then the model will assume that nu will be estimated. If
nu is fixed, you should provide the value of nu.}

\item{dim}{the dimension. Should only be provided if an
INLA mesh is not provided.}
}
\value{
index for rSPDE models.
}
\description{
Create index for INLA-based rSPDE models
}
