% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde_intrinsic.R
\name{rspde.matern.intrinsic}
\alias{rspde.matern.intrinsic}
\alias{rspde.intrinsic.matern}
\title{Intrinsic Matern rSPDE model object for INLA}
\usage{
rspde.intrinsic.matern(
  mesh,
  alpha = 2,
  prior.lkappa.mean = NULL,
  prior.ltau.mean = 1,
  prior.lkappa.prec = 0.1,
  prior.ltau.prec = 0.1,
  start.ltau = NULL,
  start.lkappa = NULL,
  debug = FALSE,
  shared_lib = "detect",
  diagonal = 1e-05,
  eigen.version = FALSE,
  mean.correction = FALSE,
  true.scaling = FALSE,
  ...
)
}
\arguments{
\item{mesh}{The mesh to build the model. It can be an \code{inla.mesh} or
an \code{inla.mesh.1d} object. Otherwise, should be a list containing elements d, the dimension, C, the mass matrix,
and G, the stiffness matrix.}

\item{alpha}{Smoothness parameter, need to be 1 or 2.}

\item{start.ltau}{Starting value for log of tau.}

\item{start.lkappa}{Starting value for log of kappa.}

\item{debug}{INLA debug argument}

\item{shared_lib}{Which shared lib to use for the cgeneric implementation?
If "detect", it will check if the shared lib exists locally, in which case it will
use it. Otherwise it will use INLA's shared library.
If "INLA", it will use the shared lib from INLA's installation. If 'rSPDE', then
it will use the local installation (does not work if your installation is from CRAN).
Otherwise, you can directly supply the path of the .so (or .dll) file.}

\item{diagonal}{Value of diagonal correction for INLA stability}

\item{eigen.version}{Use Eigen implementation? This is needed if either mean.correction or true.scaling are used.}

\item{mean.correction}{Add mean correction for extreme value models?}

\item{true.scaling}{Compute the true normalizing constant manually?}

\item{...}{Only being used internally.}

\item{prior.kappa}{a \code{list} containing the elements \code{meanlog} and
\code{sdlog}, that is, the mean and standard deviation on the log scale.}

\item{prior.tau}{a list containing the elements \code{meanlog} and
\code{sdlog}, that is, the mean and standard deviation on the log scale.}

\item{prior.kappa.mean}{Prior kappa to be used for the priors and for the starting values.}

\item{prior.tau.mean}{Prior tau to be used for the priors and for the starting values.}
}
\value{
An INLA model.
}
\description{
Creates an INLA object for a stationary intrinsic Matern model.
Currently, alpha is fixed to 2 and beta is fixed to 1.
}
