% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{rspde.matern.precision}
\alias{rspde.matern.precision}
\title{Precision matrix of the covariance-based rational approximation}
\usage{
rspde.matern.precision(kappa, nu, tau, rspde_order, dim, fem_mesh_matrices)
}
\arguments{
\item{kappa}{Range parameter of the latent process.}

\item{nu}{The smoothness parameter}

\item{tau}{The precision parameter}

\item{rspde_order}{The order of the rational approximation}

\item{dim}{The dimension}

\item{fem_mesh_matrices}{A list containing the FEM-related matrices. The list should contain elements c0, g1, g2, g3, etc.}
}
\value{
The precision matrix
}
\description{
Computes the precision matrix for the covariance-based rational SPDE
}
