% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{rspde.matern}
\alias{rspde.matern}
\title{Create INLA-based rSPDE models}
\usage{
rspde.matern(
  mesh,
  nu_upper_bound = 4,
  rspde_order = 2,
  nu = NULL,
  sharp = TRUE,
  debug = FALSE,
  optimize = TRUE,
  prior.kappa = NULL,
  prior.nu = NULL,
  prior.tau = NULL,
  start.lkappa = NULL,
  start.nu = NULL,
  start.ltau = NULL,
  prior.nu.dist = c("beta", "lognormal"),
  nu.prec.inc = 1
)
}
\arguments{
\item{mesh}{An INLA mesh}

\item{nu_upper_bound}{Upper bound for the smoothness parameter.}

\item{rspde_order}{The order of the covariance-based rational SPDE approach.}

\item{nu}{If nu is set to a parameter, nu will be kept fixed and will not be estimated.
If nu is NULL, it will be estimated.}

\item{sharp}{The sparsity graph should have the correct sparsity (costs more to perform
a sparsity analysis) or an upper bound for the sparsity? If TRUE, the graph
will have the correct sparsity.}

\item{debug}{INLA debug argument.}

\item{optimize}{Should the model be optimized? In this case the sparsities of
the matrices will be analyzed.}

\item{prior.kappa}{a list containing the elements meanlog and sdlog, that is,
the mean and standard deviation on the log scale.}

\item{prior.nu}{a list containing the elements mean and prec for beta distribution,
or meanlog and sdlog for a truncated lognormal distribution. meanlog stands for
the mean on the log scale. prec stands for the precision of a beta distribution.
sdlog stands for the standard deviation on the log scale. See help for further details.
The default precision is 50.}

\item{prior.tau}{a list containing the elements meanlog and sdlog, that is,
the mean and standard deviation on the log scale.}

\item{start.lkappa}{Starting value for log of kappa.}

\item{start.nu}{Starting value for nu.}

\item{start.ltau}{Starting value for log of tau.}

\item{prior.nu.dist}{The distribution of the smoothness parameter. The current
options are "beta" or "lognormal". The default is "beta".}

\item{nu.prec.inc}{Amount to increase the precision in the beta prior distribution.}
}
\value{
An INLA model.
}
\description{
Create INLA-based rSPDE models with general smoothness
}
