% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde_1d.R
\name{rspde.matern1d}
\alias{rspde.matern1d}
\title{Matern rSPDE model object for INLA}
\usage{
rspde.matern1d(
  loc,
  nu.upper.bound = NULL,
  rspde.order = 1,
  nu = NULL,
  parameterization = c("spde", "matern", "matern2"),
  start.nu = NULL,
  start.theta = NULL,
  prior.nu = NULL,
  theta.prior.mean = NULL,
  theta.prior.prec = 0.1,
  prior.std.dev.nominal = 1,
  prior.range.nominal = NULL,
  prior.kappa.mean = NULL,
  prior.tau.mean = NULL,
  start.lstd.dev = NULL,
  start.lrange = NULL,
  start.ltau = NULL,
  start.lkappa = NULL,
  prior.theta.param = c("theta", "spde"),
  prior.nu.dist = c("beta", "lognormal"),
  nu.prec.inc = 1,
  type.rational.approx = c("brasil", "chebfun", "chebfunLB"),
  debug = FALSE,
  shared_lib = "detect",
  ...
)
}
\arguments{
\item{loc}{A vector of spatial locations.}

\item{nu.upper.bound}{Upper bound for the smoothness parameter. If \code{NULL}, it will be set to 2.}

\item{rspde.order}{The order of the covariance-based rational SPDE approach. The default order is 1.}

\item{nu}{If nu is set to a parameter, nu will be kept fixed and will not
be estimated. If nu is \code{NULL}, it will be estimated.}

\item{parameterization}{Which parameterization to use? \code{matern} uses range, std. deviation and nu (smoothness). \code{spde} uses kappa, tau and nu (smoothness). \code{matern2} uses range-like (1/kappa), variance and nu (smoothness). The default is \code{spde}.}

\item{start.nu}{Starting value for nu.}

\item{start.theta}{Starting values for the model parameters. In the stationary case, if \code{parameterization='matern'}, then \code{theta[1]} is the std.dev and \code{theta[2]} is the range parameter.
If \code{parameterization = 'spde'}, then \code{theta[1]} is \code{tau} and \code{theta[2]} is \code{kappa}.}

\item{prior.nu}{a list containing the elements \code{mean} and \code{prec}
for beta distribution, or \code{loglocation} and \code{logscale} for a
truncated lognormal distribution. \code{loglocation} stands for
the location parameter of the truncated lognormal distribution in the log
scale. \code{prec} stands for the precision of a beta distribution.
\code{logscale} stands for the scale of the truncated lognormal
distribution on the log scale. Check details below.}

\item{theta.prior.mean}{A vector for the mean priors of \code{theta}.}

\item{theta.prior.prec}{A precision matrix for the prior of \code{theta}.}

\item{prior.std.dev.nominal}{Prior std. deviation to be used for the priors and for the starting values.}

\item{prior.range.nominal}{Prior range to be used for the priors and for the starting values.}

\item{prior.kappa.mean}{Prior kappa to be used for the priors and for the starting values.}

\item{prior.tau.mean}{Prior tau to be used for the priors and for the starting values.}

\item{start.lstd.dev}{Starting value for log of std. deviation. Will not be used if start.ltau is non-null. Will be only used in the stationary case and if \code{parameterization = 'matern'}.}

\item{start.lrange}{Starting value for log of range. Will not be used if start.lkappa is non-null. Will be only used in the stationary case and if \code{parameterization = 'matern'}.}

\item{start.ltau}{Starting value for log of tau. Will be only used in the stationary case and if \code{parameterization = 'spde'}.}

\item{start.lkappa}{Starting value for log of kappa. Will be only used in the stationary case and if \code{parameterization = 'spde'}.}

\item{prior.theta.param}{Should the lognormal prior be on \code{theta} or on the SPDE parameters (\code{tau} and \code{kappa} on the stationary case)?}

\item{prior.nu.dist}{The distribution of the smoothness parameter.
The current options are "beta" or "lognormal". The default is "lognormal".}

\item{nu.prec.inc}{Amount to increase the precision in the beta prior
distribution.}

\item{type.rational.approx}{Which type of rational approximation
should be used? The current types are "brasil", "chebfun" or "chebfunLB".}

\item{debug}{INLA debug argument}

\item{shared_lib}{Which shared lib to use for the cgeneric implementation?
\code{"detect"} and \code{"INLA"} prefer the model compiled into the INLA binary and
fall back to a compiled local rSPDE library when the symbol is unavailable.
\code{"rSPDE"} requires the local library. An existing \code{.so} or \code{.dll} path can
also be supplied directly.}

\item{...}{Only being used internally.}

\item{prior.kappa}{a \code{list} containing the elements \code{meanlog} and
\code{sdlog}, that is, the mean and standard deviation on the log scale.}

\item{prior.tau}{a list containing the elements \code{meanlog} and
\code{sdlog}, that is, the mean and standard deviation on the log scale.}

\item{prior.range}{a \code{list} containing the elements \code{meanlog} and
\code{sdlog}, that is, the mean and standard deviation on the log scale. Will not be used if prior.kappa is non-null.}

\item{prior.std.dev}{a \code{list} containing the elements \code{meanlog} and
\code{sdlog}, that is, the mean and standard deviation on the log scale. Will not be used if prior.tau is non-null.}
}
\value{
An INLA model.
}
\description{
Creates an INLA object for a stationary Matern model with
general smoothness parameter.
}
