\name{rSPDE}
\alias{rSPDE}
\title{
  Rational approximations of fractional SPDEs.
}
\description{
  \code{rSPDE} is used for approximating fractional elliptic SPDEs
   \eqn{latex}{L^\betau(s) = W}
  by rational SPDEs
  \eqn{latex}{L_1u(s) = L_2 W}
  that can be used for statistical inference. The main function for computing the
  rational operators is \code{fractional.operators}.
}
\author{
  David Bolin \email{davidbolin@gmail.com}
}
\references{
Inser reference here.
}
