% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{rspde.result}
\alias{rspde.result}
\title{rSPDE result extraction from INLA estimation results}
\usage{
rspde.result(inla, name, rspde)
}
\arguments{
\item{inla}{An inla object.}

\item{name}{A character string with the name of the rSPDE effect in the inla formula.}

\item{rspde}{The inla.rspde object used for the effect in the inla formula.}
}
\value{
Returns a list containing posterior results.
}
\description{
Extract field and parameter values and distributions for an rspde effect from an inla result object.
}
