% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde_spacetime.R
\name{rspde.spacetime}
\alias{rspde.spacetime}
\title{Space-Time Random Fields via SPDE Approximation}
\usage{
rspde.spacetime(
  mesh_space = NULL,
  mesh_time = NULL,
  space_loc = NULL,
  time_loc = NULL,
  drift = TRUE,
  alpha,
  beta,
  prior.kappa = NULL,
  prior.sigma = NULL,
  prior.rho = NULL,
  prior.gamma = NULL,
  prior.precision = NULL,
  shared_lib = "detect",
  debug = FALSE,
  ...
)
}
\arguments{
\item{mesh_space}{Spatial mesh for the FEM approximation, or a \code{metric_graph}
object for handling models on metric graphs.}

\item{mesh_time}{Temporal mesh for the FEM approximation.}

\item{space_loc}{A vector of spatial locations for mesh nodes in 1D spatial models.
This should be provided when \code{mesh_space} is not specified.}

\item{time_loc}{A vector of temporal locations for mesh nodes. This should be
provided when \code{mesh_time} is not specified.}

\item{drift}{Logical value indicating whether the drift term should be included.
If \code{FALSE}, the drift coefficient \verb{\\rho} is set to zero.}

\item{alpha}{Integer smoothness parameter \(\alpha\).}

\item{beta}{Integer smoothness parameter \(\beta\).}

\item{prior.kappa}{A list specifying the prior for the range parameter \(\kappa\).
This list may contain two elements: \code{mean} and/or \code{precision}, both of which must
be numeric scalars (numeric vectors of length 1). The precision refers to the prior
on \code{log(kappa)}. If \code{NULL}, default values will be used.}

\item{prior.sigma}{A list specifying the prior for the variance parameter \(\sigma\).
This list may contain two elements: \code{mean} and/or \code{precision}, both of which must
be numeric scalars. The precision refers to the prior on \code{log(sigma)}. If \code{NULL},
default values will be used.}

\item{prior.rho}{A list specifying the prior for the drift coefficient \(\rho\).
This list may contain two elements: \code{mean} and/or \code{precision}, both of which must
be numeric scalars. The precision applies directly to \code{rho} without log transformation.
If \code{NULL}, default values will be used. Will not be used if \code{drift = FALSE}.}

\item{prior.gamma}{A list specifying the prior for the weight \(\gamma\) in the SPDE
operator. This list may contain two elements: \code{mean} and/or \code{precision}, both of which
must be numeric scalars. The precision refers to the prior on \code{log(gamma)}. If \code{NULL},
default values will be used.}

\item{prior.precision}{A precision matrix for \verb{log(kappa), log(sigma), log(gamma), rho}. This matrix replaces the precision
element from \code{prior.kappa}, \code{prior.sigma}, \code{prior.gamma}, and \code{prior.rho} respectively. If \code{NULL}, a diagonal precision matrix
with default values will be used.}

\item{shared_lib}{String specifying which shared library to use for the Cgeneric
implementation. Options are "detect", "INLA", or "rSPDE". You may also specify the
direct path to a .so (or .dll) file.}

\item{debug}{Logical value indicating whether to enable INLA debug mode.}

\item{...}{Additional arguments passed internally for configuration purposes.}
}
\value{
An object of class \code{inla_rspde_spacetime} representing the FEM approximation of
the space-time Gaussian random field.
}
\description{
\code{rspde.spacetime} computes a Finite Element Method (FEM) approximation of a
Gaussian random field defined as the solution to the stochastic partial
differential equation (SPDE):
\deqn{d u + \gamma(\kappa^2 + \rho\cdot\nabla - \Delta)^\alpha u = \sigma dW_C}
where \( C \) is a Whittle-Matérn covariance operator with smoothness parameter
\(\beta\) and range parameter \(\kappa\). This function is designed to handle
space-time random fields using either 1D spatial models or higher-dimensional
FEM-based approaches.
}
\examples{
library(INLA)
library(MetricGraph)
graph <- metric_graph$new()
graph$build_mesh(h = 0.1)
graph$compute_fem()

# Define the time locations
time_loc <- seq(from = 0, to = 10, length.out = 11)

# Create the model
model <- rspde.spacetime(graph = graph, 
                         time_loc = time_loc, 
                         include_drift = FALSE, 
                         alpha = 2, 
                         beta = 1)

}
