% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_rspde.R
\name{rspde_create_A}
\alias{rspde_create_A}
\title{Create A matrices for INLA-based rSPDE models}
\usage{
rspde_create_A(inla_mesh, loc, rspde_order = 2, nu = NULL)
}
\arguments{
\item{inla_mesh}{An INLA mesh}

\item{loc}{Locations}

\item{rspde_order}{The order of the covariance-based rational SPDE approach.}

\item{nu}{If nu is set to a parameter, nu will be kept fixed and will not be estimated.
If nu is NULL, it will be estimated.}
}
\value{
The A matrix for rSPDE models.
}
\description{
Create A matrices for INLA-based rSPDE models
}
