% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspde_lme.R
\name{rspde_lme}
\alias{rspde_lme}
\title{rSPDE linear mixed effects models}
\usage{
rspde_lme(
  formula,
  loc,
  data,
  model = NULL,
  repl = NULL,
  optim_method = "L-BFGS-B",
  starting_values_latent = NULL,
  start_sigma_e = NULL,
  start_nu = NULL,
  nu = NULL,
  rspde_order = NULL,
  model_matrix = TRUE,
  optim_controls = list()
)
}
\arguments{
\item{formula}{Formula object describing the relation between the response variables and the fixed effects. If the response variable is a matrix, each column of the matrix will be treated as a replicate.}

\item{loc}{A vector with the names of the columns in \code{data} that contain the observation locations, or a \code{matrix} or a \code{data.frame} containing the observation locations. If the model is of class \code{metric_graph}, the locations must be either a \code{matrix} or a \code{data.frame} with two columns, or a character vector with the names of the two columns. The first column being the number of the edge, and the second column being the normalized position on the edge. If the model is a 2d model, \code{loc} must be either a \code{matrix} or \code{data.frame} with two columns or a character vector with the name of the two columns that contain the location, the first entry corresponding to the \code{x} entry and the second corresponding to the \code{y} entry.}

\item{data}{A \code{data.frame} containing the data to be used.}

\item{model}{Either an object generated by \code{matern.operators()} or \code{spde.matern.operators()}. If \code{NULL}, a simple linear regression will be performed.}

\item{repl}{Vector indicating the replicate of each observation. If \code{NULL} it will assume there is only one replicate.}

\item{optim_method}{The method to be used with \code{optim} function.}

\item{starting_values_latent}{A vector containing the starting values for the latent model. If the latent model was generated by \code{matern.operators()}, then the starting values should be provided as a vector of the form c(tau,kappa) or c(sigma,range) depending on the parameterization. If the model was generated by \code{spde.matern.operators()}, then the starting values should be provided as a vector containing the nonstationary parameters.}

\item{start_sigma_e}{Starting value for the standard deviation of the measurament error.}

\item{start_nu}{Starting value for the smoothness parameter.}

\item{nu}{If \code{NULL}, the smoothness parameter will be estimated, otherwise the smoothness parameter will be kept fixed at the provided value.}

\item{model_matrix}{logical indicating whether the model matrix should be returned as component of the returned value.}

\item{optim_controls}{Additional controls to be passed to \code{optim} or \code{optimParallel}.}
}
\value{
A list containing the fitted model.
}
\description{
Fitting linear mixed effects model with latent Whittle-Matern models.
}
