% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime.R
\name{spacetime.operators}
\alias{spacetime.operators}
\title{Space-time random fields}
\usage{
spacetime.operators(
  mesh = NULL,
  time = NULL,
  graph = NULL,
  kappa = NULL,
  tau = NULL,
  gamma = NULL,
  rho = NULL,
  alpha = NULL,
  beta = NULL
)
}
\arguments{
\item{mesh}{Spatial mesh for FEM approximation}

\item{time}{Time points for the temporal discretization}

\item{graph}{An optional \code{metric_graph} object. Replaces \code{mesh} for models on
metric graphs.}

\item{kappa}{Positive spatial range parameter}

\item{gamma}{Temporal range parameter.}

\item{rho}{Drift parameter. Real number on metric graphs and
one-dimensional spatial domains, a vector with two number on 2d domains.}

\item{alpha}{Integer smoothness parameter alpha.}

\item{beta}{Integer smoothness parameter beta.}

\item{sigma}{Positive variance parameter}
}
\description{
\code{spacetime.operators} is used for computing a FEM approximation of a Gaussian
random field defined as a solution to the SPDE
\deqn{d u + \gamma(\kappa^2 + \rho\cdot\Nabla - \Delta)^\alpha u = \sigma dW_C.}
where C is a Whittle-Matern covariance operator with smoothness parameter
\eqn{\beta} and range parameter \eqn{\kappa}
}
