% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime.R
\name{update.spacetimeobj}
\alias{update.spacetimeobj}
\title{Update space-time operator object with new parameters}
\usage{
update.spacetimeobj(
  object,
  user_kappa = NULL,
  user_sigma = NULL,
  user_gamma = NULL,
  user_rho = NULL
)
}
\arguments{
\item{object}{Space-time object created by \code{\link[=spacetime.operators]{spacetime.operators()}}}

\item{user_kappa}{kappa value to be updated.}

\item{user_sigma}{sigma value to be updated.}

\item{user_gamma}{gamma value to be updated.}

\item{user_rho}{rho value to be updated.}
}
\description{
Update space-time operator object with new parameters
}
